pyproject.toml
[project]
name = "ai-benchmarks"
version = "0.1.0"
description = "Local Ollama model benchmark harness with Claude-as-judge"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.8",
    "rich>=13.7",
    "pyyaml>=6.0",
    "psutil>=6.0",
    "typer>=0.12",
    "sqlite-utils>=3.37",
    "numpy>=2.0",
]

[project.scripts]
bench = "bench.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bench"]
